Revert "Forenkling av innholdslogikk (#529)"
This reverts commit 99e031228a4082696628c639791cd5316ab42a91.
