+++ src/innhold/innhold-view.tsx
@@ -1,10 +1,13 @@
 import * as React from 'react';
-import './innhold.less';
 import Rad from './rad';
 import AapRad from '../komponenter/aap/aap';
+import Dialog from '../komponenter/dialog/dialog';
 import Banner from '../komponenter/banner/banner';
+import Meldekort from '../komponenter/meldekort/meldekort';
+import DittSykefravaer from '../komponenter/ditt-sykefravaer/ditt-sykefravaer';
 import OkonomiRad from '../komponenter/okonomi/okonomi-rad';
 import ReaktiveringMelding from '../komponenter/reaktivering-melding';
+import Aktivitetsplan from '../komponenter/aktivitetsplan/aktivitetsplan';
 import RessurslenkerJobbsok from '../komponenter/ressurslenker-jobbsok/ressurslenker-jobbsok';
 import Egenvurdering from '../komponenter/egenvurdering/egenvurdering';
 import Motestotte from '../komponenter/motestotte/motestotte';
@@ -12,28 +15,42 @@ import './innhold.less';
 import KrrMelding from '../komponenter/krr-melding/krr-melding';
 import IARBSMelding from '../komponenter/iarbs-melding/iarbs-melding';
 import Registrert from '../komponenter/registrert/registrert';
-import AktivitetDialogMeldekort from './aktivitet-dialog-meldekort';
-import InnholdMetrics from './innhold-metrics';
+import { BrukerInfoContext } from '../ducks/bruker-info';
 
-const InnholdView = () => {
+interface OwnProps {
+    visRessurslenker: boolean;
+    skalViseIARBSPlaster: boolean;
+}
+
+const AktivitetDialogMeldekort = () => {
+    const { erSykmeldtMedArbeidsgiver } = React.useContext(BrukerInfoContext).data;
     return (
         <>
-            <InnholdMetrics />
+            <Aktivitetsplan />
+            <div className="tokol">
+                <Dialog />
+                {erSykmeldtMedArbeidsgiver ? <DittSykefravaer /> : <Meldekort />}
+            </div>
+        </>
+    );
+};
+
+const InnholdView = ({ visRessurslenker, skalViseIARBSPlaster }: OwnProps) => {
+    return (
+        <>
             <Banner />
             <Rad>
                 <ReaktiveringMelding />
                 <KrrMelding />
                 <Registrert />
-                <IARBSMelding />
+                <IARBSMelding visPlaster={skalViseIARBSPlaster} />
                 <Egenvurdering />
                 <Motestotte />
                 <AktivitetDialogMeldekort />
             </Rad>
 
             <AapRad />
-            <Rad>
-                <RessurslenkerJobbsok />
-            </Rad>
+            <Rad>{visRessurslenker ? <RessurslenkerJobbsok /> : null}</Rad>
             <OkonomiRad />
         </>
     );
