+++ src/komponenter/hent-initial-data/autentiseringsInfoFetcher.tsx
@@ -4,9 +4,11 @@ import Feilmelding from '../feilmeldinger/feilmelding';
 import { fetchData } from '../../ducks/api-utils';
 import { contextpathDittNav, erMikrofrontend } from '../../utils/app-state-utils';
 import DataProvider from './data-provider';
+import InnholdLogikkNiva4 from '../../innhold/innhold-logikk-niva4';
+import InnholdLogikkNiva3 from '../../innhold/innhold-logikk-niva3';
 import OppfolgingBrukerregistreringProvider from './oppfolging-brukerregistrering-provider';
 import * as Autentisering from '../../ducks/autentisering';
-import InnholdView from '../../innhold/innhold-view';
+import { InnloggingsNiva } from '../../ducks/autentisering';
 
 export const AUTH_API = '/api/auth';
 
@@ -27,11 +29,15 @@ const AutentiseringsInfoFetcher = () => {
             avhengigheter={[state]}
         >
             <Autentisering.AutentiseringContext.Provider value={state}>
-                <OppfolgingBrukerregistreringProvider>
-                    <DataProvider>
-                        <InnholdView />
-                    </DataProvider>
-                </OppfolgingBrukerregistreringProvider>
+                {state.data.securityLevel === InnloggingsNiva.LEVEL_3 ? (
+                    <InnholdLogikkNiva3 />
+                ) : (
+                    <OppfolgingBrukerregistreringProvider>
+                        <DataProvider>
+                            <InnholdLogikkNiva4 />
+                        </DataProvider>
+                    </OppfolgingBrukerregistreringProvider>
+                )}
             </Autentisering.AutentiseringContext.Provider>
         </Innholdslaster>
     );
