+++ src/komponenter/iarbs-melding/iarbs-melding.tsx
@@ -3,18 +3,15 @@ import { AlertStripeAdvarsel } from 'nav-frontend-alertstriper';
 import { Normaltekst } from 'nav-frontend-typografi';
 import './iarbs-melding.less';
 import tekster from '../../tekster/tekster';
-import { seIARBSPlaster } from '../../metrics/metrics';
-import { OppfolgingContext } from '../../ducks/oppfolging';
-import { BrukerInfoContext } from '../../ducks/bruker-info';
 
-const IarbsMelding = () => {
-    const visPlaster = false; // formidlingsgruppe === Formidlingsgruppe.IARBS && registreringType === RegistreringType.ALLEREDE_REGISTRERT && rettighetsgruppe !== 'AAP';
-    const { formidlingsgruppe, servicegruppe } = React.useContext(OppfolgingContext).data;
-    const { rettighetsgruppe } = React.useContext(BrukerInfoContext).data;
+export interface Props {
+    visPlaster: boolean;
+}
+
+const iarbsMelding = (props: Props) => {
+    const { visPlaster } = props;
 
     if (!visPlaster) return null;
-
-    seIARBSPlaster(formidlingsgruppe, servicegruppe, rettighetsgruppe);
 
     return (
         <AlertStripeAdvarsel className="iarbs-melding blokk-xs">
@@ -30,4 +27,4 @@ const IarbsMelding = () => {
     );
 };
 
-export default IarbsMelding;
+export default iarbsMelding;
